% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bcfrailph.R
\name{print.bcfrailph}
\alias{print.bcfrailph}
\title{Print bcfrailph}
\usage{
\method{print}{bcfrailph}(x, ...)
}
\arguments{
\item{x}{A class \code{bcfrailph} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.bcfrailph}, with some more human-readable results from \code{bcfrailph} object.
}
\description{
Generics to print the S3 class bcfrailph.
}
\details{
Calls \code{print.bcfrailph()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(24)
simdata<-simbcfrailph(p.size=300, c.rate= c(0.3),fraildistrn=c("gamma"),frail.par=c(0.5,0.5),
bhaz.arg=list(distrn = c("weibull"),shape =c(5), scale = c(0.1)),
covar.arg=list(coefs=c(2),types = c("B"),size=c(1),prob=c(0.5)))
dataa<-simdata$data

fitbcfrailph=bcfrailph(Surv(time,censor)~ X1+frailty(PID) ,data=dataa,frail_distrn=c("gamma"))
fitbcfrailph
summary(fitbcfrailph)

}
\seealso{
\code{\link{bcfrailph}}
}
