% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.shrgamsp.R
\name{print.shrgamsp}
\alias{print.shrgamsp}
\title{Print shrgamsp}
\usage{
\method{print}{shrgamsp}(x, ...)
}
\arguments{
\item{x}{A class \code{shrgamsp} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.shrgamsp}, with some more human-readable results from \code{shrgamsp} object.
}
\description{
Generics to print the S3 class shrgamsp.
}
\details{
Calls \code{print.shrgamsp()}.
}
\note{
The summary function is currently identical to the print function.
}
\seealso{
\code{\link{bcfrailph}}
}
