% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.shrgamsp.R
\name{summary.shrgamsp}
\alias{summary.shrgamsp}
\title{Print shrgamsp}
\usage{
\method{summary}{shrgamsp}(object, ...)
}
\arguments{
\item{object}{A class \code{shrgamsp} object.}

\item{...}{ignored}
}
\value{
An object of \code{summary.shrgamsp}, with some more human-readable results from \code{shrgamsp} object.
}
\description{
Generics to print the S3 class shrgamsp.
}
\details{
Calls \code{print.shrgamsp()}.
}
\note{
The summary function is currently identical to the print function.
}
\seealso{
\code{\link{bcfrailph}}
}
