% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bcfrailph.R
\name{print.bcfrailph}
\alias{print.bcfrailph}
\title{Print bcfrailph}
\usage{
\method{print}{bcfrailph}(x, ...)
}
\arguments{
\item{x}{A class \code{bcfrailph} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.bcfrailph}, with some more human-readable results from \code{bcfrailph} object.
}
\description{
Generics to print the S3 class bcfrailph.
}
\details{
Calls \code{print.bcfrailph()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(4)
simdata<-simbcfrailph(psize=300, cenr= c(0.3),beta=c(2),frailty=c("gamma"),
frailpar=c(0.5,0.5),bhaz=c("weibull"),
bhazpar=list(shape =c(5), scale = c(0.1)),
covartype= c("B"),covarpar=list(fargs=c(1),sargs=c(0.5)))
dataa<-simdata$data

fitbcfrailph=bcfrailph(Surv(time,censor)~ X1+frailty(PID) ,data=dataa,frail_distrn=c("gamma"))
fitbcfrailph

}
\seealso{
\code{\link{bcfrailph}}
}
